Remove pesky Account Unknowns from your system
This script traverses the registry matching SIDS in HKEY_USERS against the WMI win32_userprofile class. If a sid is in the registry, but not the class table, it is an orphaned account, and will be removed from the registry. This removes the Account Unknown entry from System > User Profiles.